requirements.txt
certifi==2021.5.30
chardet==4.0.0
Django==2.2.24
idna==2.10
Pillow==8.2.0
pip==21.1.2
pytz==2021.1
requests==2.25.1
setuptools==28.8.0
sqlparse==0.4.1
urllib3==1.26.5